Implement conditional rendering for empty logs

+++ app.php
@@ -11,9 +11,13 @@
   # Getting all the accidents data
   $accidents = $DBconnection->query("SELECT * FROM accidents WHERE user_id = {$_SESSION["user"]["user_id"]}");
 
-  $search_statement = $DBconnection->query("SELECT timestamp FROM accidents WHERE user_id = {$_SESSION["user"]["user_id"]} ORDER BY timestamp DESC LIMIT 1");
-
-  $last_accident_timestamp = $search_statement->fetch(PDO::FETCH_ASSOC)["timestamp"] ?? null;
+  if ($accidents->rowCount() > 0) {
+    $search_statement = $DBconnection->query("SELECT timestamp FROM accidents WHERE user_id = {$_SESSION["user"]["user_id"]} ORDER BY timestamp DESC LIMIT 1");
+    $last_accident_timestamp = $search_statement->fetch(PDO::FETCH_ASSOC)["timestamp"] ?? null;
+  }
+  else {
+    $last_accident_timestamp = null;
+  }
 
   function level_color(int $level): string {
     return match($level) {
@@ -30,39 +34,45 @@
 
 <main class="h-full mb-16">
   <div class="flex justify-center items-center bg-gray-600 h-64">
-    <div id="counter" class="text-9xl font-mono mr-4" data-timestamp="<?= $last_accident_timestamp ?>"></div>
-    <div class="text-2xl font-sans">days without accidents</div>
+    <?php if ($accidents->rowCount() > 0): ?>
+      <div id="counter" class="text-9xl font-mono mr-4" data-timestamp="<?= $last_accident_timestamp ?>"></div>
+      <div class="text-2xl font-sans">days without accidents</div>
+    <?php endif ?>
   </div>
   <div class="container m-auto p-3">
-    <?php foreach ($accidents as $accident) : ?>
-      <div class="w-full max-w-2xl px-4 py-3 mx-auto bg-white mt-4">
-        <div class="flex items-center justify-between">
-          <span class="text-xs font-mono font-light text-gray-500"><?= $accident["timestamp"] ?></span>
-          <span class="px-3 py-1 text-xs text-white-800 uppercase bg-<?= level_color($accident["level"]) ?>-300 rounded-md">Level: <?= $accident["level"] ?></span>
-        </div>
-
-        <div>
-          <h1 class="text-xl font-semibold text-gray-800"><?= $accident["title"] ?></h1>
-          <p class="text-base text-gray-800"><?= $accident["description"] ?></p>
-        </div>
-        <?php if (isset($accident["more_info"])) : ?>
-          <div>
-            <div class="flex items-center mt-2 text-gray-700">
-              <span>More info on:</span>
-              <a class="mx-2 text-blue-600 cursor-pointer hover:underline">example.com</a>
-            </div>
+    <?php if ($accidents->rowCount() > 0): ?>
+      <?php foreach ($accidents as $accident) : ?>
+        <div class="w-full max-w-2xl px-4 py-3 mx-auto bg-white mt-4">
+          <div class="flex items-center justify-between">
+            <span class="text-xs font-mono font-light text-gray-500"><?= $accident["timestamp"] ?></span>
+            <span class="px-3 py-1 text-xs text-white-800 uppercase bg-<?= level_color($accident["level"]) ?>-300 rounded-md">Level: <?= $accident["level"] ?></span>
           </div>
-        <?php endif ?>
-        <div class="flex gap-2 text-xs mt-2">
-          <a href="delete.php?id=<?= $accident["id"] ?>" class="text-red-600 underline">
-            Delete
-          </a>
-          <a href="edit.php?id=<?= $accident["id"] ?>" class="text-gray-600 underline">
-            Edit
-          </a>
+
+          <div>
+            <h1 class="text-xl font-semibold text-gray-800"><?= $accident["title"] ?></h1>
+            <p class="text-base text-gray-800"><?= $accident["description"] ?></p>
+          </div>
+          <?php if (isset($accident["more_info"])) : ?>
+            <div>
+              <div class="flex items-center mt-2 text-gray-700">
+                <span>More info on:</span>
+                <a class="mx-2 text-blue-600 cursor-pointer hover:underline">example.com</a>
+              </div>
+            </div>
+          <?php endif ?>
+          <div class="flex gap-2 text-xs mt-2">
+            <a href="delete.php?id=<?= $accident["id"] ?>" class="text-red-600 underline">
+              Delete
+            </a>
+            <a href="edit.php?id=<?= $accident["id"] ?>" class="text-gray-600 underline">
+              Edit
+            </a>
+          </div>
         </div>
-      </div>
-    <?php endforeach ?>
+      <?php endforeach ?>
+    <?php else: ?>
+      <div class="text-center">No accidents to show!</div>
+    <?php endif ?>
   </div>
   <a href="add.php" class="m-auto w-32 justify-center border-y border-x border-black cursor-pointer flex items-center px-2 py-2 font-normal tracking-wide capitalize transition-colors duration-200 transform bg-white hover:bg-gray-100 rounded-md focus:outline-none focus:ring focus:ring-blue-300 focus:ring-opacity-80">
     <svg class="w-5 h-5 mx-1 text-black" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor">
